Renamed the application 'planetarium' to 'shows'. Fixed urls.
