--- planetarium/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from planetarium.models import ShowTheme, AstronomyShow
-
-admin.site.register(ShowTheme)
-admin.site.register(AstronomyShow)
-
