--- planetarium/models.py
@@ -1,37 +1,0 @@
-import os
-import uuid
-from django.db import models
-from django.utils.text import slugify
-
-
-class ShowTheme(models.Model):
-    name = models.CharField(max_length=255, unique=True)
-
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        ordering = ["id"]
-
-
-def astronomy_show_image_file_path(instance, filename):
-    _, extension = os.path.splitext(filename)
-    filename = f"{slugify(instance.title)}-{uuid.uuid4()}{extension}"
-
-    return os.path.join("uploads/astronomy_shows/", filename)
-
-
-class AstronomyShow(models.Model):
-    title = models.CharField(max_length=255, unique=True)
-    description = models.TextField()
-    show_themes = models.ManyToManyField(
-        ShowTheme,
-        related_name="astronomy_shows"
-    )
-    image = models.ImageField(null=True, upload_to=astronomy_show_image_file_path)
-
-    def __str__(self):
-        return self.title
-
-    class Meta:
-        ordering = ["id"]
