--- planetarium/serializers.py
@@ -1,40 +1,0 @@
-from rest_framework import serializers
-from planetarium.models import AstronomyShow, ShowTheme
-
-
-class ShowThemeSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ShowTheme
-        fields = ("id", "name")
-
-
-class AstronomyShowSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = AstronomyShow
-        fields = ("id", "title", "description", "show_themes")
-
-
-class AstronomyShowListSerializer(AstronomyShowSerializer):
-    show_themes = serializers.SlugRelatedField(
-        many=True,
-        read_only=True,
-        slug_field="name"
-    )
-
-    class Meta:
-        model = AstronomyShow
-        fields = ("id", "title", "description", "show_themes", "image")
-
-
-class AstronomyShowDetailSerializer(AstronomyShowSerializer):
-    show_themes = ShowThemeSerializer(many=True, read_only=True)
-
-    class Meta:
-        model = AstronomyShow
-        fields = ("id", "title", "description", "show_themes", "image")
-
-
-class AstronomyShowImageSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = AstronomyShow
-        fields = ("id", "image")
