--- planetarium/views.py
@@ -1,106 +1,0 @@
-from drf_spectacular.types import OpenApiTypes
-from drf_spectacular.utils import extend_schema, OpenApiParameter
-from rest_framework import mixins, status
-from rest_framework.decorators import action
-from rest_framework.permissions import IsAdminUser
-from rest_framework.response import Response
-from rest_framework.viewsets import GenericViewSet
-from planetarium.models import ShowTheme, AstronomyShow
-from planetarium.permissions import IsAdminOrIfAuthenticatedReadOnly
-from planetarium.serializers import ShowThemeSerializer, AstronomyShowSerializer, AstronomyShowListSerializer, \
-    AstronomyShowDetailSerializer, AstronomyShowImageSerializer
-
-
-class ShowThemeViewSet(
-    mixins.CreateModelMixin,
-    mixins.ListModelMixin,
-    mixins.DestroyModelMixin,
-    GenericViewSet,
-):
-    queryset = ShowTheme.objects.all()
-    serializer_class = ShowThemeSerializer
-    permission_classes = (IsAdminOrIfAuthenticatedReadOnly,)
-
-
-class AstronomyShowViewSet(
-    mixins.ListModelMixin,
-    mixins.CreateModelMixin,
-    mixins.RetrieveModelMixin,
-    mixins.DestroyModelMixin,
-    mixins.UpdateModelMixin,
-    GenericViewSet,
-):
-    queryset = AstronomyShow.objects.prefetch_related("show_themes")
-    serializer_class = AstronomyShowSerializer
-    permission_classes = (IsAdminOrIfAuthenticatedReadOnly,)
-
-    @staticmethod
-    def _params_to_ints(qs):
-        """Converts a list of string IDs to a list of integers"""
-        return [int(str_id) for str_id in qs.split(",")]
-
-    def get_queryset(self):
-        """Retrieve the astronomy_shows with filters"""
-        title = self.request.query_params.get("title")
-        show_themes = self.request.query_params.get("show_themes")
-
-        queryset = self.queryset
-
-        if title:
-            queryset = queryset.filter(title__icontains=title)
-
-        if show_themes:
-            show_themes_ids = self._params_to_ints(show_themes)
-            queryset = queryset.filter(show_themes__id__in=show_themes_ids)
-
-        return queryset.distinct()
-
-    def get_serializer_class(self):
-        if self.action == "list":
-            return AstronomyShowListSerializer
-
-        if self.action == "retrieve":
-            return AstronomyShowDetailSerializer
-
-        if self.action == "upload_image":
-            return AstronomyShowImageSerializer
-
-        return AstronomyShowSerializer
-
-    @action(
-        methods=["POST"],
-        detail=True,
-        url_path="upload-image",
-        permission_classes=[IsAdminUser],
-    )
-    def upload_image(self, request, pk=None):
-        """Endpoint for uploading image to specific astronomy show"""
-        astronomy_show = self.get_object()
-        serializer = self.get_serializer(astronomy_show, data=request.data)
-
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
-
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    # Only for documentation purposes
-    @extend_schema(
-        parameters=[
-            OpenApiParameter(
-                "show_themes",
-                type={"type": "list", "items": {"type": "number"}},
-                description="Filter by show theme id (ex. ?show_themes=2,5)",
-            ),
-            OpenApiParameter(
-                "title",
-                type=OpenApiTypes.STR,
-                description=(
-                    "Filter by astronomy show title "
-                    "(ex. ?title=stars)"
-                ),
-            ),
-        ]
-    )
-    def list(self, request, *args, **kwargs):
-        return super().list(request, *args, **kwargs)
