+++ shows/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from shows.models import ShowTheme, AstronomyShow
+
+admin.site.register(ShowTheme)
+admin.site.register(AstronomyShow)
+
